requirements.txt
click==8.1.3
colorama==0.4.4
Flask==2.1.2
Flask-PyMongo==2.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pymongo==4.1.1
python-decouple==3.6
python-dotenv==0.20.0
Werkzeug==2.1.2
